extract a common distributed testing class for ut test=develop test=document_preview
